+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change log
 ==========
 
 * 🎁 Add "Update" / "Help" menu entries
+* 🎁 Add "Find ICAO airport codes" menu entries
 
 1.1.0
 -----
@@ -11,7 +12,7 @@ Change log
 1.0.1
 -----
 
-* 💊 Improve handling of ICAO codes
+* 💊 Improve handling of ICAO airport codes
 * 💊 Fix behaviour of seconds in UTC time
 
 1.0.0